refactor CommandParser into individual Parsers for each command
