+++ src/main/java/seedu/task/logic/parser/commands/DeleteCommandParser.java
@@ -1,0 +1,30 @@
+package seedu.task.logic.parser.commands;
+
+import static seedu.task.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
+
+import java.util.Optional;
+
+import seedu.task.logic.commands.Command;
+import seedu.task.logic.commands.DeleteCommand;
+import seedu.task.logic.commands.IncorrectCommand;
+import seedu.task.logic.parser.CommandParser;
+
+// @@author A0147944U
+public class DeleteCommandParser {
+
+    /**
+     * Parses arguments in the context of the delete task command.
+     *
+     * @param args full command args string
+     * @return the prepared command
+     */
+    public static Command prepareDelete(String args) {
+
+        Optional<Integer> index = CommandParser.parseIndex(args);
+        if (!index.isPresent()) {
+            return new IncorrectCommand(String.format(MESSAGE_INVALID_COMMAND_FORMAT, DeleteCommand.MESSAGE_USAGE));
+        }
+
+        return new DeleteCommand(index.get());
+    }
+}
