+++ src/main/java/seedu/task/logic/parser/commands/FindCommandParser.java
@@ -1,0 +1,39 @@
+package seedu.task.logic.parser.commands;
+
+import static seedu.task.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
+
+import java.util.Arrays;
+import java.util.HashSet;
+import java.util.Set;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+
+import seedu.task.logic.commands.Command;
+import seedu.task.logic.commands.FindCommand;
+import seedu.task.logic.commands.IncorrectCommand;
+
+// @@author A0147944U
+public class FindCommandParser {
+
+    // One or more keywords separated by whitespace
+    private static final Pattern KEYWORDS_ARGS_FORMAT = Pattern.compile("(?<keywords>\\S+(?:\\s+\\S+)*)");
+
+    /**
+     * Parses arguments in the context of the find task command.
+     *
+     * @param args
+     *            full command args string
+     * @return the prepared command
+     */
+    public static Command prepareFind(String args) {
+        final Matcher matcher = KEYWORDS_ARGS_FORMAT.matcher(args.trim());
+        if (!matcher.matches()) {
+            return new IncorrectCommand(String.format(MESSAGE_INVALID_COMMAND_FORMAT, FindCommand.MESSAGE_USAGE));
+        }
+
+        // keywords delimited by whitespace
+        final String[] keywords = matcher.group("keywords").split("\\s+");
+        final Set<String> keywordSet = new HashSet<>(Arrays.asList(keywords));
+        return new FindCommand(keywordSet);
+    }
+}
