+++ src/main/java/seedu/task/logic/parser/commands/SelectCommandParser.java
@@ -1,0 +1,29 @@
+package seedu.task.logic.parser.commands;
+
+import static seedu.task.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
+
+import java.util.Optional;
+
+import seedu.task.logic.commands.Command;
+import seedu.task.logic.commands.IncorrectCommand;
+import seedu.task.logic.commands.SelectCommand;
+import seedu.task.logic.parser.CommandParser;
+
+// @@author A0147944U
+public class SelectCommandParser {
+
+    /**
+     * Parses arguments in the context of the select task command.
+     *
+     * @param args full command args string
+     * @return the prepared command
+     */
+    public static Command prepareSelect(String args) {
+        Optional<Integer> index = CommandParser.parseIndex(args);
+        if (!index.isPresent()) {
+            return new IncorrectCommand(String.format(MESSAGE_INVALID_COMMAND_FORMAT, SelectCommand.MESSAGE_USAGE));
+        }
+
+        return new SelectCommand(index.get());
+    }
+}
